Update lightgallery.ts
This is related to #993. If lightgallery is imported in a script that is used in `<head>` rather than in `<body>`, `settings.container` will be null, as the settings are initialized on import of the script not when lightgallery is called. This PR proposes a change to make it work as long as the dev wraps their script in some kind of DOMReady / window.load.

I do not know if this is the very best place to add but it has to be before or in `buildStructure`.